requirements.txt
pygame-ce==2.5.3
mypy==1.15.0
ruff==0.11.11
